MueLu/HHG: need numLevels of hierarchy
After creating the hierarchy, we need to now its number of levels in
order to run the HHG-specific V-cycle. For now, we extract this
information from the HHG prolongator data structure.

Part of #4084.

Note: also fix Matlab syntax. Use ' instead of " and move LID widget to Utils file
