+++ packages/muelu/research/mmayr/composite_to_regions/src/main.cpp
@@ -80,12 +80,14 @@
 #include <Teuchos_Assert.hpp>
 #include <Teuchos_Comm.hpp>
 #include <Teuchos_CommandLineProcessor.hpp>
+#include <Teuchos_DefaultComm.hpp>
 #include <Teuchos_FancyOStream.hpp>
 #include <Teuchos_ScalarTraits.hpp>
 #include <Teuchos_XMLParameterListHelpers.hpp>
 
 
 #include <SetupRegionHierarchy_def.hpp>
+#include <HHG_Utils_def.hpp>
 
 /* This is a driver that is not included in any other file.
  * So, we should be fine to create useful typedefs for Xpetra here and use them in the entire file.
@@ -121,27 +123,6 @@ enum InputDataIndices
   inpData_cornerY,
   inpData_nGhosts,
   inpData_firstLIDsOfGhosts
-};
-
-// this little widget handles application specific data
-// used to implement LIDregion()
-struct widget {
-  int maxRegPerProc;
-  int *minGIDComp;
-  int *maxGIDComp;
-  int *myRegions;
-  Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> *colMap;
-  int maxRegPerGID;
-  Teuchos::RCP<MultiVector> regionsPerGIDWithGhosts;
-  int *gDim, *lDim, *lowInd;
-  int *trueCornerx; // global coordinates of region
-  int *trueCornery; // corner within entire 2D mesh
-  int *relcornerx;  // coordinates of corner relative
-  int *relcornery;  // to region corner
-  int *lDimx;
-  int *lDimy;
-  int nx;
-  int myRank;
 };
 
 //! Print an object in regional layout to screen
@@ -818,6 +799,9 @@ int main(int argc, char *argv[]) {
   // Run V-cycle
   {
     std::cout << myRank << " | Running V-cycle ..." << std::endl;
+
+    // Extract the number of levels from the prolongator data structure
+    numLevels = regProlong.size();
 
     TEUCHOS_TEST_FOR_EXCEPT_MSG(!(numLevels>0), "We require numLevel > 0. Probably, numLevel has not been set, yet.");
